Export_csvcomponentgrades is a plugin that exports grades from an assignment for the following grading methods, Rubric and Marking Guide, into a CSV file.
Export_csvcomponentgrades is a plugin that exports grades from an assignment for the following grading methods, Rubric and Marking Guide, into a CSV file.

Once installed, as a user who has permission to grade the assignment, navigate to an assignment that uses a Rubric or a Marking Guide grading method. In the Assignment Administration menu, a new menu item will appear: "Export Marking Guide Grades (.csv)" or "Export Rubric Grades (.csv)" as appropriate; click it to download a CSV file containing that assignment's component grades for all students in the course. Note: the download is NOT an Excel spreadsheet.

In the CSV file, the information exported is both the students’ information (first name, last name, username and student ID) together with their respective grades.